rename `Post` folder and file to `Posts`
