+++ src/pages/matik/index.tsx
@@ -1,7 +1,7 @@
 import {NextPage} from 'next'
 
 import {PageLayout} from '@/components/PageLayout/PageLayout'
-import {Posts} from '@/components/Post/Post'
+import {Posts} from '@/components/Posts/Posts'
 
 const Matik: NextPage = () => (
   <PageLayout title="Seminár Matik">
